+++ database/migrations/2022_06_30_094538_update_version_110.php
@@ -3,6 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Modules\Core\Models\NotificationPush;
 
 return new class extends Migration
 {
@@ -28,6 +29,21 @@ return new class extends Migration
                 $table->tinyInteger('is_virtual')->nullable()->default(0);
             }
         });
+
+
+        Schema::table(NotificationPush::getTableName(),function (Blueprint $table){
+            if(!Schema::hasColumn(NotificationPush::getTableName(),'for_admin')){
+                $table->boolean('for_admin',30)->default(0)->nullable();
+            }
+        });
+        Schema::table('product_variations',function (Blueprint $table){
+            if(!Schema::hasColumn('product_variations','variation_type')){
+                $table->tinyInteger('variation_type')->nullable()->default(0);
+                $table->bigInteger('vendor_id')->nullable();
+                $table->bigInteger('parent_id')->nullable();
+            }
+        });
+
     }
 
     /**