+++ themes/Base/Views/global/message.blade.php
@@ -6,7 +6,7 @@
 @endif
 
 
-@if ($message = Session::get('error'))
+@if ($message = Session::get('error',Session::get('danger')))
     <div class="alert alert-danger alert-dismissible fade show">
         {{ $message }}
         <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>